types.h clean up

See merge request tkochi/m88clone2!23

+++ src/common/soundsrc.h
@@ -2,14 +2,14 @@
 
 #pragma once
 
-#include "win32/types.h"
+#include <stdint.h>
 
 typedef int16_t Sample;
 
 class SoundSource {
  public:
   virtual int Get(Sample* dest, int size) = 0;
-  virtual ulong GetRate() = 0;
+  virtual uint32_t GetRate() = 0;
   virtual int GetChannels() = 0;
   virtual int GetAvail() = 0;
 };
@@ -19,7 +19,7 @@ typedef int32_t SampleL;
 class SoundSourceL {
  public:
   virtual int Get(SampleL* dest, int size) = 0;
-  virtual ulong GetRate() = 0;
+  virtual uint32_t GetRate() = 0;
   virtual int GetChannels() = 0;
   virtual int GetAvail() = 0;
 };
